* origin/master: (307 commits)
  BUG: refs #0458. Changed check to search for first existing parent dir.
  ENH: refs #446. Server side quota validation for uploading new revisions
  ENH: refs #446. Client side validation for upload quota on revision upload
  BUG: refs #0458. Added is_writable check to prevent breaking json parsing.
  ENH: refs #0446. Refactor upload client code
  BUG: refs #0456. Fixes issue where the metadata flowed out of its div.
  ENH: refs #442. Resign applet jar with new certificate. Updated related docs.
  ENH: refs #0448. Add a test for the web api info reporting methods
  ENH: refs #448. Add web api info methods to report enables modules and methods
  STYLE: refs #0445. Userpage Actions links should have mouseover highlighting
  ENH: refs #0445. Show a user's community invitations on their user page
  ENH: refs #445. Private user should be allowed to view their own userpage
  BUG: refs #0445. Alphabet goes all the way to Z
  STYLE: refs #0340. Fix trailing whitespace, and fix stub admin ctrlr test
  ENH: refs #0340. Add stub admin ctrlr test for better coverage reporting
  ENH: refs #340 #444. Add a test for admin setting fuctionality
  ENH: refs #340 #444. Allow admin status to be set on a user by an admin user
  BUG: refs #340 #444. Don't change first/last name display when updating profile
  ENH: refs #0340. Fix test so it won't break other tests
  STYLE: refs #0340. Fix UI look and feel for delete user dialog
  ...

+++ core/models/base/BitstreamModelBase.php
@@ -1,13 +1,21 @@
 <?php
 /*=========================================================================
-MIDAS Server
-Copyright (c) Kitware SAS. 20 rue de la Villette. All rights reserved.
-69328 Lyon, FRANCE.
+ MIDAS Server
+ Copyright (c) Kitware SAS. 26 rue Louis Guérin. 69100 Villeurbanne, FRANCE
+ All rights reserved.
+ More information http://www.kitware.com
 
-See Copyright.txt for details.
-This software is distributed WITHOUT ANY WARRANTY; without even
-the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-PURPOSE.  See the above copyright notices for more information.
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+         http://www.apache.org/licenses/LICENSE-2.0.txt
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
 =========================================================================*/
 
 /** Bitstream Base Model*/
@@ -51,13 +59,14 @@ abstract class BitstreamModelBase extends AppModel
     {
     if(!$bitstream instanceof BitstreamDao)
       {
-      throw new Zend_Exception("Error param.");
+      throw new Zend_Exception('Must pass a bitstream dao');
       }
     $checksum = $bitstream->getChecksum();
     $path = $bitstream->getFullPath();
     $assetstore = $bitstream->getAssetstore();
     parent::delete($bitstream);
-    if($assetstore->getType() != MIDAS_ASSETSTORE_REMOTE && $this->getByChecksum($checksum) == false)
+    if(file_exists($path) && $assetstore->getType() != MIDAS_ASSETSTORE_REMOTE
+       && $this->getByChecksum($checksum) == false)
       {
       unlink($path);
       }
